# accounting/openapi/openapi.json
{
...
  "host": "127.0.0.1:45881",
  "info": {
    "title": "Komunitin accounting API",
    "version": "0.0.1",
    "license": {
      "name": "AGPL-3.0-only"
    },
    "description": "API for the Komunitin payments system. This documentation is automatically generated from the serivce test suite."
  },
  "paths": {
...
    "/{code}/currency/settings": {
      "get": {
        "summary": "/{code}/currency/settings",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": [
          "currency"
        ]
      },
      "patch": {
        "summary": "/{code}/currency/settings",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "type": "string",
            "example": "TEST"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "attributes": {
                      "type": "object",
                      "properties": {
                        "defaultOnPaymentCreditLimit": {
                          "type": "number",
                          "example": 1300
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "currency-settings"
                    },
                    "id": {
                      "type": "string",
                      "example": "f3eeabe3-2772-49ae-be20-437a066f9a38"
                    },
                    "attributes": {
                      "type": "object",
                      "properties": {
                        "defaultAllowPayments": {
                          "type": "boolean",
                          "example": true
                        },
                        "enableExternalPayments": {
                          "type": "boolean",
                          "example": true
                        },
                        "defaultAllowTagPayments": {
                          "type": "boolean",
                          "example": false
                        },
                        "defaultInitialCreditLimit": {
                          "type": "number",
                          "example": 1000
                        },
                        "externalTraderCreditLimit": {
                          "type": "number",
                          "example": 1000
                        },
                        "defaultAcceptPaymentsAfter": {
                          "type": "number",
                          "example": 1209600
                        },
                        "defaultAllowPaymentRequests": {
                          "type": "boolean",
                          "example": true
                        },
                        "defaultOnPaymentCreditLimit": {
                          "type": "number",
                          "example": 1300
                        },
                        "defaultAllowExternalPayments": {
                          "type": "boolean",
                          "example": true
                        },
                        "enableExternalPaymentRequests": {
                          "type": "boolean",
                          "example": false
                        },
                        "defaultAcceptPaymentsWhitelist": {
                          "type": "array",
                          "items": {},
                          "example": [
                            null
                          ]
                        },
                        "defaultAllowTagPaymentRequests": {
                          "type": "boolean",
                          "example": false
                        },
                        "defaultAcceptPaymentsAutomatically": {
                          "type": "boolean",
                          "example": false
                        },
                        "defaultAllowExternalPaymentRequests": {
                          "type": "boolean",
                          "example": false
                        },
                        "defaultAcceptExternalPaymentsAutomatically": {
                          "type": "boolean",
                          "example": false
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string"
                      },
                      "code": {
                        "type": "string"
                      },
                      "title": {
                        "type": "string"
                      },
                      "detail": {
                        "type": "string"
                      }
                    }
                  },
                  "example": [
                    {
                      "status": "400",
                      "code": "BadRequest",
                      "title": "Bad Request",
                      "detail": ""
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string"
                      },
                      "code": {
                        "type": "string"
                      },
                      "title": {
                        "type": "string"
                      },
                      "detail": {
                        "type": "string"
                      }
                    }
                  },
                  "example": [
                    {
                      "status": "403",
                      "code": "Forbidden",
                      "title": "Forbidden",
                      "detail": "User not found in currency TES2"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string"
                      },
                      "code": {
                        "type": "string"
                      },
                      "title": {
                        "type": "string"
                      },
                      "detail": {
                        "type": "string"
                      }
                    }
                  },
                  "example": [
                    {
                      "status": "404",
                      "code": "NotFound",
                      "title": "Not Found",
                      "detail": "Currency with code TES2 not found"
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string"
                      },
                      "code": {
                        "type": "string"
                      },
                      "title": {
                        "type": "string"
                      },
                      "detail": {
                        "type": "string"
                      }
                    }
                  },
                  "example": [
                    {
                      "status": "500",
                      "code": "NotImplemented",
                      "title": "Not Implemented",
                      "detail": "Change the external trader credit limit not implemented yet"
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "currency"
        ],
        "produces": [
          "application/vnd.api+json"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      }
    },
...
    "/{code}/accounts.csv": {
      "get": {
        "summary": "/{code}/accounts.csv",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string",
              "example": "id,created,updated,code,status,balance,creditLimit,maximumBalance,key,user.id,transfers.2026,transfers.2025,transfers.2024\n92a1307d-860b-4f72-b2f6-e1b2414cdd5d,2026-06-29T14:53:49.419Z,2026-06-29T14:53:54.401Z,TEST0001,active,0.00,0.10,10.00,GDFD7M4ZKAOONB654DEX5GHO7ZVOMJAJ7U4ARU3FRJJ2H4FASNTENKOL,2,0,2,2\n636b4296-e84b-4289-b170-27ca96b15779,2026-06-29T14:53:48.415Z,2026-06-29T14:53:53.519Z,TEST1000,active,0.00,1000000.00,,GBTO3XGX6VA4A7C46ZUKX3GWK55ZC3NIBHYWYHCCQFKFYRQ56XNZYY5M,1,0,2,2\nf7f38257-de38-4e49-83ac-41089af6a9bf,2026-06-29T14:53:50.427Z,2026-06-29T14:53:50.427Z,TEST2000,active,0.00,0.20,2.00,GCAJTMWBR34V432D5SF57MJCLY3M6OHEIV4VSNFJG3BQXGFQ5VBVATAA,1,0,0,0"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string"
                      },
                      "code": {
                        "type": "string"
                      },
                      "title": {
                        "type": "string"
                      },
                      "detail": {
                        "type": "string"
                      }
                    }
                  },
                  "example": [
                    {
                      "status": "400",
                      "code": "BadRequest",
                      "title": "Bad Request",
                      "detail": "Invalid account transfer stats field transfers.ytd"
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "accounts"
        ],
        "produces": [
          "text/csv",
          "application/vnd.api+json"
        ]
      }
    },
...
    "/{code}/transfers.csv": {
      "get": {
        "summary": "/{code}/transfers.csv",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": [
          "transfers"
        ]
      }
    },
...
    "/{code}/trustlines/sync": {
      "post": {
        "summary": "/{code}/trustlines/sync",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": [
          "trustlines"
        ]
      }
    },
...
    "/{code}/stats/amount": {
      "get": {
        "summary": "/{code}/stats/amount",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "stats"
                    },
                    "attributes": {
                      "type": "object",
                      "properties": {
                        "values": {
                          "type": "array",
                          "items": {
                            "type": "number"
                          },
                          "example": [
                            1868700
                          ]
                        },
                        "from": {
                          "type": "string",
                          "example": "2024-06-01T00:00:00.000Z"
                        },
                        "to": {
                          "type": "string",
                          "example": "2024-07-01T00:00:00.000Z"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [],
        "produces": [
          "application/vnd.api+json"
        ]
      }
    },
    "/{code}/stats/accounts": {
      "get": {
        "summary": "/{code}/stats/accounts",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": [
          "accounts"
        ]
      }
    },
    "/{code}/stats/transfers": {
      "get": {
        "summary": "/{code}/stats/transfers",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": [
          "transfers"
        ]
      }
    },
...
  },
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "authorization": {
      "name": "Authorization",
      "in": "header",
      "type": "apiKey",
      "description": "Bearer JWT from the configured Auth issuer for audience urn:komunitin:api. Use accounting:read for reads, accounting:write for writes, or superadmin on admin operations."
    }
  },
  "servers": [
    {
      "url": "https://accounting.komunitin.org",
      "description": "Production server"
    },
    {
      "url": "https://accounting.demo.komunitin.org",
      "description": "Demo server"
    },
    {
      "url": "http://localhost:2025",
      "description": "Local development server"
    }
  ],
  "swagger": "2.0",
...
}